split up scenarios in multiple files as lazy-loading scenarios grow
