+++ src/scenarios/README.md
@@ -1,0 +1,1 @@
+scenarios contains the high-level playbook for the test suite
