+++ src/usecases/README.md
@@ -1,0 +1,2 @@
+use cases contains the detailed DOM interactions to perform a given use case, may also do some assertions.
+use cases are often used in multiple scenarios.
